Close EPM session after each request
TIA portal uses EPM requests for identifying devices on the network.

Previously the sessions were not closed, resulting in "Out of session resources"
when a number of requests have been done.

Profinet uses EPM in a limited way. If we for some reason we need more advanced
EPM handling in the future, implement a timeout for each EPM session.

Closes #396