Increase default number of alarms
To properly handle incoming alarm frames the receive thread must have high enough priority.
Then it is possible that many alarm frames are queued before the stack has a chance to
handle them. Adapt to the burst of 6 alarm frames in the Alarm APMS test case for the
ART Tester.